LayoutTests/imported/w3c/web-platform-tests/cookies/samesite/sandbox-iframe-nested.https-expected.txt



FAIL SameSite cookies with intervening sandboxed iframe and navigation assert_equals: `samesite_lax=0.07283376622792892` in request to `https://localhost:9443`. expected (undefined) undefined but got (string) "0.07283376622792892"
